requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scipy==1.13.1
joblib==1.4.2
requests==2.31.0
python-dotenv==1.0.0
anthropic>=0.25.0
openai>=1.0.0
gunicorn==21.2.0
